labels, mnemonics, print, tests
A symbol table, labels and EQU were added. All mnemonics are now assembled although only a few
are actually simulated. Print without argument, prints the LSB of the register as an ascii character.
Print with an argument of C'd' print the whole register as a decimal integer. Print with an
argument of C'xy' prints the two ascii characters in the small string, ignoring the register.